W1 HW: Milestone #3
Create account / Sign in

At the last welcome screen, the user can choose to Sign Up or Sign In.
In the Dropbox app, after leaving the Welcome screens, they can't
return. They are taken immediately to the Create Account or Sign In
form. If they hit the back button, they are taken to another signed out
screen, shown below. Sign In Flow
In the diagram above, I have two copies of the signed out page which I
called CreateAccountRootViewController and SignInRootViewController. As
soon as these view controller load, they immediately segue into the
Create Account or Sign In form, respectively.
To accomplish this, after creating a segue in storyboard, you can
trigger the transition manually from within the viewDidLoad() method of
the CreateAccountRootViewController and SignInRootViewController. This
will push the view controller onto the navigation stack.
Q. "How will I ever go back to this screen if the segue is performed
every time the view controller loads?"
A. The viewDidLoad() method is only called when a view controller is
loaded. Using Segues, a view controller is loaded each time it is
segued to, but the view controller where the segue originated is not
loaded upon return from the segue.
In order to return from the CreateAccountViewController and
SignInViewController we need to add a button over the "< Welcome"
button and create an action. We can programmatically return to the
previous view controller by popping the view controller.
I added in a few intermediate screens to simulate filling out the
create account and sign in forms.