W1_HW Milestone #4
Terms of Service

At the end of the create account sequence, after clicking "create",
there is a screen image that shows an action with options to view
"Terms of Service" or "Agree". Create another view controller and
create a modal segue to that view controller from a button placed over
the "terms of service" action sheet button. In the new view controller,
add an image view containing the terms of service image.
Add a button over the "done", create an action and add code to
programmatically dismiss the view controller and return to the previous
view controller.