W1 HW1: Milestone #5
Signed In, New User

Add a Tab Bar Controller and configure it with 4 view controllers.
From the CreateAccountViewController screen, create a modal segue to
the tab bar controller.
For each view controller in the tab, embed in Navigation Controller.
Add the images for the new user screens in the respective view
controllers. The settings image should be embedded within a scroll view.