Milestone 2: Basic Layout and Transition
Basic Layout and Transition
In the Storyboard, add the 5 UIImageViews to the feed, remember to
enable user interaction for each image view.
Add tap gesture recognizers to each UIImageView.
Create a PhotoViewController that has an image view embedded in a
scroll view and a Done button.
Create a modal segue from the FeedViewController to the
PhotoViewController and give it a unique identifier.
Attach an action to each gesture recognizer (they can be connected to
the same action method). In the action, perform the segue. Guide:
Performing Manual Segue.
Set the image of the PhotoViewController. Guide: Passing Data in Segues
Connect the Done button to an action that dismisses the view controller.